[chore] : #3 패키지 이름 screen -> ui로 변경
