+++ app/src/main/java/org/sopt/and/presentation/ui/MainActivity.kt
@@ -1,10 +1,9 @@
-package org.sopt.and.presentation.screen
+package org.sopt.and.presentation.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
